Move files to predicate folder
