+++ src/main/java/jimmy/mcgymmy/model/predicate/NameContainsKeywordsPredicate.java
@@ -1,9 +1,10 @@
-package jimmy.mcgymmy.model.food;
+package jimmy.mcgymmy.model.predicate;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import jimmy.mcgymmy.commons.util.StringUtil;
+import jimmy.mcgymmy.model.food.Food;
 
 /**
  * Tests that a {@code Food}'s {@code Name} matches any of the keywords given.
