+++ src/test/java/jimmy/mcgymmy/logic/commands/FindCommandTest.java
@@ -14,7 +14,7 @@ import jimmy.mcgymmy.logic.parser.CommandParserTestUtil;
 import jimmy.mcgymmy.model.Model;
 import jimmy.mcgymmy.model.ModelManager;
 import jimmy.mcgymmy.model.UserPrefs;
-import jimmy.mcgymmy.model.food.NameContainsKeywordsPredicate;
+import jimmy.mcgymmy.model.predicate.NameContainsKeywordsPredicate;
 import jimmy.mcgymmy.testutil.TypicalFoods;
 
 /**
